Add follows endpoint
